requirements.txt
pyyaml==6.0.1
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.4
tqdm==4.66.1
tiktoken==0.5.2 
